refactor: migrate money amounts to Decimal/Numeric across domain and …

+++ app/models/asset_valuation.py
@@ -1,6 +1,7 @@
 from datetime import datetime
+from decimal import Decimal
 
-from sqlalchemy import DateTime, Float, ForeignKey, Integer, String, Text, func
+from sqlalchemy import DateTime, ForeignKey, Integer, Numeric, String, Text, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.database import Base
@@ -14,7 +15,7 @@ class AssetValuation(Base):
         Integer, ForeignKey("accounts.id"), nullable=False, index=True
     )
     date: Mapped[datetime] = mapped_column(DateTime, nullable=False, index=True)
-    value: Mapped[float] = mapped_column(Float, nullable=False)
+    value: Mapped[Decimal] = mapped_column(Numeric(18, 2), nullable=False)
     currency: Mapped[str] = mapped_column(String(10), default="USD")
     source: Mapped[str] = mapped_column(String(50), default="manual")
     notes: Mapped[str | None] = mapped_column(Text)
